+++ Assets/OxGKit/ActionSystem/Scripts/Runtime/DefaultActions/ParallelDelayAction.cs
@@ -46,7 +46,7 @@ namespace OxGKit.ActionSystem
         {
             this.UpdateParallelActions(dt);
 
-            if (this.IsAllDone())
+            if (this._IsAllDone())
             {
                 this.MarkAsDone();
                 this.ClearActions();
@@ -83,16 +83,16 @@ namespace OxGKit.ActionSystem
         {
             foreach (var action in this._queueRunningActions.ToArray())
             {
-                if (action.IsDone()) return;
+                if (action.IsAllDone()) return;
                 action.RunUpdate(dt);
             }
         }
 
-        protected bool IsAllDone()
+        private bool _IsAllDone()
         {
             foreach (var action in this._queueRunningActions.ToArray())
             {
-                if (!action.IsDone()) return false;
+                if (!action.IsAllDone()) return false;
             }
 
             return true;