+++ Assets/OxGKit/ActionSystem/Scripts/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## CHANGELOG
 
+## [0.0.2-preview] - 2023-05-21=8
+- Fixed OnDone invoke timing (If call MarkAsDone will invoke OnDone).
+- Modified done flags are IsDone and IsAllDone. 
+
 ## [0.0.1-preview] - 2023-05-21
 - Added ActionSystem.